+++ words.txt
@@ -1,9 +1,9 @@
 hello|привет|0
 dog|собака|0
 cat|кошка|0
-man|мужчина|0
+man|мужчина|1
 horse|лошадь|0
 bag|сумка|0
 mouse|мышь|0
-cave|пещера|2
-river|река|1
+cave|пещера|3
+river|река|2
